Adding GET functionality

+++ src/main/java/com/example/docvisitorregister/repository/PatientQueryRepository.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface PatientQueryRepository {
 
-    List<Object[]> getPatientsLastVisit(int pageSize, int pageNo, String firstName, String lastName, List<Integer> doctorIds);
+    List<Object[]> getPatientsLastVisit(int pageSize, int pageNo, String firstName, String lastName, List<Long> doctorIds);
 }
